+++ aiohomematic/model/custom/climate.py
@@ -351,15 +351,15 @@ class BaseCustomDpClimate(CustomDataPoint):
     async def set_profile(self, profile: ClimateProfile, collector: CallParameterCollector | None = None) -> None:
         """Set new profile."""
 
-    @inspector()
+    @inspector
     async def enable_away_mode_by_calendar(self, start: datetime, end: datetime, away_temperature: float) -> None:
         """Enable the away mode by calendar on thermostat."""
 
-    @inspector()
+    @inspector
     async def enable_away_mode_by_duration(self, hours: int, away_temperature: float) -> None:
         """Enable the away mode by duration on thermostat."""
 
-    @inspector()
+    @inspector
     async def disable_away_mode(self) -> None:
         """Disable the away mode on thermostat."""
 
@@ -375,7 +375,7 @@ class BaseCustomDpClimate(CustomDataPoint):
             return True
         return super().is_state_change(**kwargs)
 
-    @inspector()
+    @inspector
     async def copy_schedule(self, target_climate_data_point: BaseCustomDpClimate) -> None:
         """Copy schedule to target device."""
 
@@ -388,7 +388,7 @@ class BaseCustomDpClimate(CustomDataPoint):
             values=raw_schedule,
         )
 
-    @inspector()
+    @inspector
     async def copy_schedule_profile(
         self,
         source_profile: ScheduleProfile,
@@ -418,7 +418,7 @@ class BaseCustomDpClimate(CustomDataPoint):
             do_validate=False,
         )
 
-    @inspector()
+    @inspector
     async def get_schedule_profile(self, profile: ScheduleProfile) -> PROFILE_DICT:
         """Return a schedule by climate profile."""
         if not self._supports_schedule:
@@ -426,7 +426,7 @@ class BaseCustomDpClimate(CustomDataPoint):
         schedule_data = await self._get_schedule_profile(profile=profile)
         return schedule_data.get(profile, {})
 
-    @inspector()
+    @inspector
     async def get_schedule_profile_weekday(self, profile: ScheduleProfile, weekday: ScheduleWeekday) -> WEEKDAY_DICT:
         """Return a schedule by climate profile."""
         if not self._supports_schedule:
@@ -478,7 +478,7 @@ class BaseCustomDpClimate(CustomDataPoint):
 
         return schedule_data
 
-    @inspector()
+    @inspector
     async def set_schedule_profile(
         self, profile: ScheduleProfile, profile_data: PROFILE_DICT, do_validate: bool = True
     ) -> None:
@@ -518,7 +518,7 @@ class BaseCustomDpClimate(CustomDataPoint):
             values=_get_raw_schedule_paramset(schedule_data=schedule_data),
         )
 
-    @inspector()
+    @inspector
     async def set_simple_schedule_profile(
         self,
         profile: ScheduleProfile,
@@ -531,7 +531,7 @@ class BaseCustomDpClimate(CustomDataPoint):
         )
         await self.set_schedule_profile(profile=profile, profile_data=profile_data)
 
-    @inspector()
+    @inspector
     async def set_schedule_profile_weekday(
         self,
         profile: ScheduleProfile,
@@ -559,7 +559,7 @@ class BaseCustomDpClimate(CustomDataPoint):
             values=_get_raw_schedule_paramset(schedule_data=schedule_data),
         )
 
-    @inspector()
+    @inspector
     async def set_simple_schedule_profile_weekday(
         self,
         profile: ScheduleProfile,
@@ -844,7 +844,7 @@ class CustomDpRfThermostat(BaseCustomDpClimate):
                     value=_HM_WEEK_PROFILE_POINTERS_TO_NAMES[profile_idx], collector=collector
                 )
 
-    @inspector()
+    @inspector
     async def enable_away_mode_by_calendar(self, start: datetime, end: datetime, away_temperature: float) -> None:
         """Enable the away mode by calendar on thermostat."""
         await self._client.set_value(
@@ -854,14 +854,14 @@ class CustomDpRfThermostat(BaseCustomDpClimate):
             value=_party_mode_code(start=start, end=end, away_temperature=away_temperature),
         )
 
-    @inspector()
+    @inspector
     async def enable_away_mode_by_duration(self, hours: int, away_temperature: float) -> None:
         """Enable the away mode by duration on thermostat."""
         start = datetime.now() - timedelta(minutes=10)
         end = datetime.now() + timedelta(hours=hours)
         await self.enable_away_mode_by_calendar(start=start, end=end, away_temperature=away_temperature)
 
-    @inspector()
+    @inspector
     async def disable_away_mode(self) -> None:
         """Disable the away mode on thermostat."""
         start = datetime.now() - timedelta(hours=11)
@@ -1095,7 +1095,7 @@ class CustomDpIpThermostat(BaseCustomDpClimate):
             if profile_idx := self._profiles.get(profile):
                 await self._dp_active_profile.send_value(value=profile_idx, collector=collector)
 
-    @inspector()
+    @inspector
     async def enable_away_mode_by_calendar(self, start: datetime, end: datetime, away_temperature: float) -> None:
         """Enable the away mode by calendar on thermostat."""
         await self._client.put_paramset(
@@ -1109,14 +1109,14 @@ class CustomDpIpThermostat(BaseCustomDpClimate):
             },
         )
 
-    @inspector()
+    @inspector
     async def enable_away_mode_by_duration(self, hours: int, away_temperature: float) -> None:
         """Enable the away mode by duration on thermostat."""
         start = datetime.now() - timedelta(minutes=10)
         end = datetime.now() + timedelta(hours=hours)
         await self.enable_away_mode_by_calendar(start=start, end=end, away_temperature=away_temperature)
 
-    @inspector()
+    @inspector
     async def disable_away_mode(self) -> None:
         """Disable the away mode on thermostat."""
         await self._client.put_paramset(