+++ aiohomematic/model/device.py
@@ -427,14 +427,14 @@ class Device(ContextMixin, PayloadMixin):
         elif self._channel_groups[channel_no] != group_no:
             return
 
-    @inspector()
+    @inspector
     async def create_central_links(self) -> None:
         """Create a central links to support press events on all channels with click events."""
         if self.relevant_for_central_link_management:  # pylint: disable=using-constant-test
             for channel in self._channels.values():
                 await channel.create_central_link()
 
-    @inspector()
+    @inspector
     async def remove_central_links(self) -> None:
         """Remove central links."""
         if self.relevant_for_central_link_management:  # pylint: disable=using-constant-test
@@ -584,7 +584,7 @@ class Device(ContextMixin, PayloadMixin):
             for dp in self.generic_data_points:
                 dp.fire_data_point_updated_callback()
 
-    @inspector()
+    @inspector
     async def export_device_definition(self) -> None:
         """Export the device definition for current device."""
         try:
@@ -613,7 +613,7 @@ class Device(ContextMixin, PayloadMixin):
             for callback_handler in self._firmware_update_callbacks:
                 callback_handler()
 
-    @inspector()
+    @inspector
     async def update_firmware(self, refresh_after_update_intervals: tuple[int, ...]) -> bool:
         """Update the firmware of the homematic device."""
         update_result = await self._client.update_device_firmware(device_address=self._address)
@@ -628,7 +628,7 @@ class Device(ContextMixin, PayloadMixin):
 
         return update_result
 
-    @inspector()
+    @inspector
     async def load_value_cache(self) -> None:
         """Init the parameter cache."""
         if len(self.generic_data_points) > 0:
@@ -640,7 +640,7 @@ class Device(ContextMixin, PayloadMixin):
             self._address,
         )
 
-    @inspector()
+    @inspector
     async def reload_paramset_descriptions(self) -> None:
         """Reload paramset for device."""
         for (
@@ -884,7 +884,7 @@ class Channel(ContextMixin, PayloadMixin):
         """Return the unique_id of the channel."""
         return self._unique_id
 
-    @inspector()
+    @inspector
     async def create_central_link(self) -> None:
         """Create a central link to support press events."""
         if self._has_key_press_events and not await self._has_central_link():
@@ -892,7 +892,7 @@ class Channel(ContextMixin, PayloadMixin):
                 address=self._address, value_id=REPORT_VALUE_USAGE_VALUE_ID, ref_counter=1
             )
 
-    @inspector()
+    @inspector
     async def remove_central_link(self) -> None:
         """Remove a central link."""
         if self._has_key_press_events and await self._has_central_link() and not await self._has_program_ids():
@@ -900,7 +900,7 @@ class Channel(ContextMixin, PayloadMixin):
                 address=self._address, value_id=REPORT_VALUE_USAGE_VALUE_ID, ref_counter=0
             )
 
-    @inspector()
+    @inspector
     async def cleanup_central_link_metadata(self) -> None:
         """Cleanup the metadata for central links."""
         if metadata := await self._device.client.get_metadata(address=self._address, data_id=REPORT_VALUE_USAGE_DATA):
@@ -1260,7 +1260,7 @@ class _DefinitionExporter:
         self._device_address: Final = device.address
         self._random_id: Final[str] = f"VCU{int(random.randint(1000000, 9999999))}"
 
-    @inspector()
+    @inspector
     async def export_data(self) -> None:
         """Export data."""
         device_descriptions: Mapping[str, DeviceDescription] = (