+++ aiohomematic/model/generic/__init__.py
@@ -100,7 +100,7 @@ _SWITCH_DP_TO_SENSOR: Final[Mapping[str | tuple[str, ...], Parameter]] = {
 }
 
 
-@inspector()
+@inspector
 def create_data_point_and_append_to_channel(
     channel: hmd.Channel,
     paramset_key: ParamsetKey,