+++ aiohomematic/model/hub/switch.py
@@ -31,13 +31,13 @@ class ProgramDpSwitch(GenericProgramDataPoint):
         """Get the value of the data_point."""
         return self._is_active
 
-    @inspector()
+    @inspector
     async def turn_on(self) -> None:
         """Turn the program on."""
         await self.central.set_program_state(pid=self._pid, state=True)
         await self._central.fetch_program_data(scheduled=False)
 
-    @inspector()
+    @inspector
     async def turn_off(self) -> None:
         """Turn the program off."""
         await self.central.set_program_state(pid=self._pid, state=False)